+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Optional inline SVG workaround ([#1](https://github.com/marp-team/marpit/pull/1))
+
 ## v0.0.0 - 2018-03-24
 
 * Initial release. It does not cover integration test and document fully.